api/requirements.txt
# requirements.txt
flask==2.2.3
flask-cors==3.0.10
psycopg2-binary==2.9.5
werkzeug==2.2.2
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.0.0
python-dotenv==1.0.0
itsdangerous==2.1.2